Added getRecordAsXml that returns an record as XML.
In order for this to work specify the xml option in a record. 
The xml option is an array that has 2 keys. 

ignore_fields is an array of fields that are not displayed in the xml
include_relations(optional) is an array of relations that should be included in
the xml.