Add models for testing arbitrary object hydration.
The model consists of two entities: Customer and Address. The models are linked
to one another through a one to one relationship that is owned by the customer.

There third object represents the combined data for both models. This
CustomerAddressView object will be used to test the arbitrary object hydration.