removed erroneous tip
I'm working through the tutorial myself and got to this point.

Don’t forget to add a require_once call for this class to the bootstrap.php

This advice is wrong. The class is located in the src directory and so is autoloaded when required. No require statement is necessary.