Fix for no dot on Class Names and Schema Names
If you work with PostgreSQL Schemas, You should filter the names of table to generate Correct name for PHP Classes. This way allow not write dots (.) as part of the Class Name.

In addition we have added two properties in which data from a table schema and namespace of a class are stored, this is useful when working with PostgreSQL Schemas.

Additionally, there is a variable ($schema) that must be contained in the class "Doctrine\DBAL\Schema\Table" on an $schema possible property but this is not available. Or can be the value of "name" property of the Class "Doctrine\DBAL\Schema\SchemaConfig". (This last is recomended)

This is a fix because Doctrine entityGenerate proccess is generating the name of PHP classes with dots (.) when you work with schemas on PostgreSQL. The Name of Classes on PHP can not have dot on its Name. Therefore Doctrine2 can't generate PHP Classes with dots on its names.

These data are also placed in the metadata.