Initial work of nullable embedded objects.
There's a lot of things to be improved (mostly on UnitOfWork), but it's fully working with basic tests.
The idea was to have a simple (and clean) way to override the nullable property of embedded objects, so I've
added the nullable attribute on the Embedded annotation and it have 3 possible values:

- NULL: The nullable option that was defined on the attributes of the embeddable class won't be overriden;
- TRUE: All attributes of the embeddable class will be marked as nullable;
- FALSE: All attributes of the embeddable class will be marked as non-nullable.

The usage is quite simple and can be seen at the ValueObjectsTest::testCRUDOfNullableEmbedded() test case.