[DDC-3582] Fix hydration of nested embeddables
The wrong class is chosen when hydrating embeddables that are part of a nested
structure. See DDC3582Test for a demonstration. The fix is to use the 'class'
instead of the 'embeddedClass' to instantiate the embeddable in
ReflectionEmbeddedProperty.

This causes one of the scenarios in the test for ReflectionEmbeddedProperty to
fail. But as far as I can see, this scenario wouldn't occur in reality anyway,
because an abstract embeddable is always extended in a real world mapping.