Fix OneToOne unique constraint mapping
When two independent Entites use the same property name to associate the
OneToOne-ness the ClassMetadataInfo will generate a unique constraint
based on the property name and databases like SQLite will fail because
the indexes need to be unique in the schema space.

This hack prepends the owning table name to the constraint, so instead
of foo_uniq it will become tablename_foo_uniq.

Ideally the NamingStrategy should be consulted but the current API does
not have any support for this.