Use AS part of the query from platform
This change was part of the "Add SAP Adaptive Server Enterprise vendor support" PR doctrine/doctrine2#5752. It has been seperated because it is not ASE specific and would increase platform independence in general.

It depends on the  "Add selectAliasColumn to AbstractPlatform" PR doctrine/dbal#2359.

It is required in the platform if the "AS" keyword may differ in platforms (which is probably not very common) and for creating workarounds when using deferred tables to simulate ROWNUM() by identity columns.