ObjectHydrator: fixed initialization of left joined collections
There is initialization of collections which are considered to be empty and are not fetched twice from commit https://github.com/doctrine/doctrine2/commit/3407620bf80ec5d09475dcab1f0e089aecfcc064.

But implementation is wrong, it decided collection to be empty from first row only. If there were data in next row collection was already initialized. So it all ended with empty related collections marked as initialized.