Add quoted column keys while hydrating column information.
We can't hydrate column information while using a quote strategy because all or the most internal doctrine mappings use the quoted column alias name.

While hydrating a result row data we know we've only got the unquoted column alias and there is no way to quote an existing column alias.
So we need to split the column alias back to the real column name and the alias counter.
This is easy because we know it's get generated in the format '<column_name>_<counter>' by the basic entity persister.

With this information we can use the current quote strategy to generate a quoted column alias.