Fix CommitOrderCalculator::visit()
This fixes a commit order when persisting many entities with cyclic
associations within one flush. Before the fix, when multiple cycles
were presented in the precedence graph and more than two classes
existed in within a cycle the calculated commit order was likely to be
wrong, resulting in not null constraint violation when saving to
a database. An example of such situation is provided in
CommitOrderCalculatorTest::testCommitOrdering4() - this is a test case
with won’t pass without the introduced fix.

Prior to the fix the weight of an edge in a class precedence graph was
compared only to the weight of the subsequent edges. Therefore, in case
of cycles containing more than two classes the result was
unpredictable. The fixed code uses also the weight of further nodes in
the precedence graph. It is achieved by adding minimum weight argument
to the visitor function of CommitOrderCalulator – edges having weight
below the minimum weight are ignored. The minimum weight is increased
each time a node is revisited, i. e. a cycle is detected. It guarantees
that edges of highest priority representing not nullable relationships
are processed first.