Introduce `toGenerator()` on AbstractQuery
The point is to have both `toIterable()` and `toGenerator()` so consumers are able to rely on stricter type.

So if they choose `toGenerator()` in order to rely on `Generator` return type, they can assume it is not "rewindable", eg. they can analyze in code that the Generator value is not iterated more than once