Fix CommitOrderCalculator::addDependency()
It may happen that an entity class A has multiple many-to-one
associations with the same target entity B. These associations may
differ in “nullable” parameter. In such situation
CommitOrderCalculator::addDependency() is called multiple times for the
same pair of $fromHash and $toHash, with different value of $weight
argument. Before the fix the subsequent call to
CommitOrderCalculator::addDependency() for the same $fromHash and
$toHash overrode $weight set at the previous call. It caused
unpredictable commit order of entities which resulted in not null
constraint violation when saving to a database.

Fixed code preserves the highest value of weight argument. If at
subsequent call the weight is lower than previously established weight
the addition of edge to the dependency list is skipped.