Implement expression to filter some field with IN a native temporary table.
This patch is to fix a specific issue in a large project and is
submitted to doctrine/orm for proof of concept/idea. This is not
intended to be a final solution as I want some input from the doctrine
developers on how to approch this problem (which will be described
later). Also... gotta find a better name than "MEMBERTEMP {NOT} OF"

In my project I have a very complicated query that 1)
includes a large section with joins and subqueryies for ACL and 2) a
large IN expression with over 8000 int IDs. This query is written with
DQL and uses over 1 minute to execute which is unexceptable. I have been
researching alot and discovered that mysql is not performing the query
efficiently as I had hoped, but putting all the IDs inside a temporary
table and then later referencing that table in a IN-subquery makes the
query use about 0.60 sec. DQL however does not allow someone to
reference native tables as DQL is specificly designed (in my eyes) to
work with entities known to Doctrine. I have tried several
workarounds or "hacks" to allow referencing (for instance using a
"FakeEntity" where I set the table name in runtime on ClassMetadata) the
temporary table. Using a custom string function does not work and using
a custom SQL walker seem to not be a feasable solution as we still cant
reference the table name itself without an entity.

The following patch tries to replicate the "MEMBER OF" expression but
instead allow a "string primary" for a collection. This string is
never "validated" and is just put directly into the resulting SQL as a
table name. No criteria support has been added since we put all the ids
already filtered into the temporary table.